+++ release-notes.rst
@@ -1,7 +1,20 @@
 Release notes for chevah.empirical
 ==================================
 
-0.10.2 - 23/09/2013
+
+0.10.4 - 17/09/2013
+-------------------
+
+* Wait 10 seconds 2nd API call for getting a group.
+
+
+0.10.3 - 17/09/2013
+-------------------
+
+* Wait 5 seconds for 2nd API call for getting a group.
+
+
+0.10.2 - 16/09/2013
 -------------------
 
 * Try 2 different API calls to wait for group creation.